requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
clickhouse-connect==0.7.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pydantic==2.5.3
python-dotenv==1.0.1
